Integer division now returns a double

+++ plugins/org.eclipse.epsilon.eol.engine/src/org/eclipse/epsilon/eol/types/NumberUtil.java
@@ -94,7 +94,7 @@ public class NumberUtil {
 			return n1.longValue() / n2.longValue();
 		}
 		else {
-			return n1.intValue() / n2.intValue();
+			return ((double)n1.intValue()) / n2.intValue();
 		}
 	}
